Fix for sending notifications with include_external_user_ids
I was having an issue when using include_external_user_ids; The OneSignal API was returning an error: `Platforms You may only send to one delivery channel at a time. Make sure you are only including one of push platforms, isEmail, or isSms.`

This is fixed by setting channel_for_external_user_ids to Push, so the API knows which players to send.

This is a quick fix, but could possibly be made configurable.